# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 知行创境教育创作平台 API
  version: 1.0.0
  description: |
    基于《知行创境》教育创作平台 V2.0 PRD 整理的后端接口 OpenAPI 文档。
    基础路径为 `/api/v1`，认证方式为 Bearer Token。
servers:
  - url: /api/v1
    description: Default API Server
...
paths:
...
  /users/me/profile:
    put:
      tags: [Auth]
      summary: 修改个人资料
      operationId: updateMyProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /users/me/password:
    put:
      tags: [Auth]
      summary: 修改密码
      operationId: updateMyPassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordRequest'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /tasks:
    get:
      tags: [Tasks]
      summary: 获取任务列表
      operationId: getTasks
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/TaskStatus'
        - in: query
          name: classId
          schema:
            type: integer
            format: int64
        - in: query
          name: keyword
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 任务分页列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PagedTaskSummary'
        default:
          $ref: '#/components/responses/ErrorResponse'
    post:
      tags: [Tasks]
      summary: 创建任务
      operationId: createTask
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TaskDetail'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /tasks/{taskId}:
    get:
      tags: [Tasks]
      summary: 获取任务详情
      operationId: getTaskDetail
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: 任务详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TaskDetail'
        default:
          $ref: '#/components/responses/ErrorResponse'
    put:
      tags: [Tasks]
      summary: 更新任务
      operationId: updateTask
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /tasks/{taskId}/excellent-exhibitions:
    get:
      tags: [Tasks]
      summary: 获取任务优秀作品
      operationId: getExcellentExhibitionsByTask
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: 优秀作品列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ExhibitionSummary'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /tasks/{taskId}/progress:
    get:
      tags: [Tasks]
      summary: 教师查看任务进度
      operationId: getTaskProgress
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: 任务进度
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TaskProgress'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /tasks/{taskId}/submit:
    post:
      tags: [Submissions]
      summary: 提交任务作品
      operationId: submitTaskWork
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubmissionRequest'
      responses:
        '200':
          description: 提交成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SubmissionDetail'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /tasks/{taskId}/submissions:
    get:
      tags: [Submissions]
      summary: 获取任务提交记录
      operationId: getTaskSubmissions
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: 提交记录列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/SubmissionDetail'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /submissions/{submissionId}:
    get:
      tags: [Submissions]
      summary: 获取提交详情
      operationId: getSubmissionDetail
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SubmissionId'
      responses:
        '200':
          description: 提交详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SubmissionDetail'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /submissions/{submissionId}/reviews:
    post:
      tags: [Submissions]
      summary: 教师评分与点评
      operationId: createSubmissionReview
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SubmissionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubmissionReviewRequest'
      responses:
        '200':
          description: 点评成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SubmissionReview'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /users/{userId}/homepage:
    get:
      tags: [Users]
      summary: 获取个人主页
      operationId: getUserHomepage
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: 个人主页
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserHomepage'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /users/me/portfolio:
    get:
      tags: [Users]
      summary: 获取我的作品集
      operationId: getMyPortfolio
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 作品集列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/PortfolioItem'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /notifications:
    get:
      tags: [Notifications]
      summary: 获取通知列表
      operationId: getNotifications
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: readStatus
          schema:
            type: string
            enum: [unread, read]
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 通知分页列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PagedNotification'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /notifications/read:
    post:
      tags: [Notifications]
      summary: 标记通知已读
      operationId: readNotifications
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReadNotificationsRequest'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      in: query
      name: pageSize
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    TaskId:
      in: path
      name: taskId
      required: true
      schema:
        type: integer
        format: int64
...
    UserId:
      in: path
      name: userId
      required: true
      schema:
        type: integer
        format: int64
...
    SubmissionId:
      in: path
      name: submissionId
      required: true
      schema:
        type: integer
        format: int64
  responses:
    SuccessResponse:
      description: 操作成功
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            code: 0
            message: ok
            data: {}
            requestId: 4f9b6e24e2ef4f4f9d6c
    ErrorResponse:
      description: 业务异常
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                code: 40100
                message: 未登录或 token 失效
                data: null
                requestId: 4f9b6e24e2ef4f4f9d6c
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
        data:
          nullable: true
        requestId:
          type: string
      required: [code, message, requestId]
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 40001
        message:
          type: string
          example: 参数错误
        data:
          nullable: true
        requestId:
          type: string
      required: [code, message, requestId]
    UserRole:
      type: string
      enum: [student, teacher, parent, admin, visitor]
    TaskStatus:
      type: string
      enum: [draft, published, closed, archived]
    ExhibitionStatus:
      type: string
      enum: [draft, published, archived]
    ExhibitionVisibility:
      type: string
      enum: [private, class, public]
    SubmissionStatus:
      type: string
      enum: [submitted, reviewed, returned]
...
    ClassInfo:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        grade:
          type: string
        academicYear:
          type: string
    SimpleUser:
      type: object
      properties:
        id:
          type: integer
          format: int64
        role:
          $ref: '#/components/schemas/UserRole'
        name:
          type: string
        nickname:
          type: string
        avatarUrl:
          type: string
          format: uri
    UserProfile:
      type: object
      properties:
        id:
          type: integer
          format: int64
        role:
          $ref: '#/components/schemas/UserRole'
        realName:
          type: string
        nickname:
          type: string
        avatarUrl:
          type: string
          format: uri
        email:
          type: string
          format: email
        mobile:
          type: string
        bio:
          type: string
        school:
          $ref: '#/components/schemas/SchoolInfo'
        classInfo:
          $ref: '#/components/schemas/ClassInfo'
...
    UpdateProfileRequest:
      type: object
      properties:
        nickname:
          type: string
        avatarUrl:
          type: string
          format: uri
        bio:
          type: string
    UpdatePasswordRequest:
      type: object
      required: [oldPassword, newPassword]
      properties:
        oldPassword:
          type: string
          format: password
        newPassword:
          type: string
          format: password
...
    TaskMaterial:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        materialType:
          type: string
          enum: [image, video, audio, link, document]
        url:
          type: string
          format: uri
        description:
          type: string
    TaskSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        coverUrl:
          type: string
          format: uri
        description:
          type: string
        teacher:
          $ref: '#/components/schemas/SimpleUser'
        startTime:
          type: string
          format: date-time
        dueTime:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/TaskStatus'
    TaskDetail:
      allOf:
        - $ref: '#/components/schemas/TaskSummary'
        - type: object
          properties:
            evaluationCriteria:
              type: string
            backgroundMaterials:
              type: array
              items:
                $ref: '#/components/schemas/TaskMaterial'
            targetClasses:
              type: array
              items:
                $ref: '#/components/schemas/ClassInfo'
            excellentExhibitions:
              type: array
              items:
                $ref: '#/components/schemas/ExhibitionSummary'
    CreateTaskRequest:
      type: object
      required: [title, description, dueTime, targetClassIds]
      properties:
        title:
          type: string
        coverUrl:
          type: string
          format: uri
        description:
          type: string
        backgroundMaterials:
          type: array
          items:
            $ref: '#/components/schemas/TaskMaterial'
        evaluationCriteria:
          type: string
        startTime:
          type: string
          format: date-time
        dueTime:
          type: string
          format: date-time
        targetClassIds:
          type: array
          items:
            type: integer
            format: int64
    UpdateTaskRequest:
      allOf:
        - $ref: '#/components/schemas/CreateTaskRequest'
    TaskProgressGroup:
      type: object
      properties:
        exhibitionId:
          type: integer
          format: int64
        groupName:
          type: string
        leaderName:
          type: string
        memberCount:
          type: integer
        progressPercent:
          type: integer
        submissionStatus:
          $ref: '#/components/schemas/SubmissionStatus'
        updatedAt:
          type: string
          format: date-time
    TaskProgress:
      type: object
      properties:
        taskId:
          type: integer
          format: int64
        submittedCount:
          type: integer
        reviewedCount:
          type: integer
        groups:
          type: array
          items:
            $ref: '#/components/schemas/TaskProgressGroup'
    ExhibitionStats:
      type: object
      properties:
        viewCount:
          type: integer
        likeCount:
          type: integer
        favoriteCount:
          type: integer
        commentCount:
          type: integer
    ExhibitionSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        taskId:
          type: integer
          format: int64
        title:
          type: string
        coverUrl:
          type: string
          format: uri
        summary:
          type: string
        status:
          $ref: '#/components/schemas/ExhibitionStatus'
        visibility:
          $ref: '#/components/schemas/ExhibitionVisibility'
        groupName:
          type: string
        ownerId:
          type: integer
          format: int64
        author:
          $ref: '#/components/schemas/SimpleUser'
        latestVersionNo:
          type: integer
        publishedVersionNo:
          type: integer
        stats:
          $ref: '#/components/schemas/ExhibitionStats'
        tags:
          type: array
          items:
            type: string
...
    CreateSubmissionRequest:
      type: object
      required: [exhibitionId]
      properties:
        exhibitionId:
          type: integer
          format: int64
        submitRemark:
          type: string
    SubmissionReview:
      type: object
      properties:
        id:
          type: integer
          format: int64
        reviewer:
          $ref: '#/components/schemas/SimpleUser'
        score:
          type: number
          format: double
        commentText:
          type: string
        commentAudioUrl:
          type: string
          format: uri
        isPublic:
          type: boolean
        createdAt:
          type: string
          format: date-time
    SubmissionDetail:
      type: object
      properties:
        id:
          type: integer
          format: int64
        taskId:
          type: integer
          format: int64
        exhibitionId:
          type: integer
          format: int64
        submitter:
          $ref: '#/components/schemas/SimpleUser'
        versionNo:
          type: integer
        submitRemark:
          type: string
        submissionStatus:
          $ref: '#/components/schemas/SubmissionStatus'
        submittedAt:
          type: string
          format: date-time
        reviewedAt:
          type: string
          format: date-time
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/SubmissionReview'
    CreateSubmissionReviewRequest:
      type: object
      properties:
        score:
          type: number
          format: double
        commentText:
          type: string
        commentAudioUrl:
          type: string
          format: uri
        isPublic:
          type: boolean
          default: true
...
    UserHomepageStats:
      type: object
      properties:
        exhibitionCount:
          type: integer
        likeCount:
          type: integer
        teacherPraiseCount:
          type: integer
    PortfolioItem:
      type: object
      properties:
        exhibitionId:
          type: integer
          format: int64
        title:
          type: string
        coverUrl:
          type: string
          format: uri
        publishedAt:
          type: string
          format: date-time
        likeCount:
          type: integer
    UserHomepage:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserProfile'
        stats:
          $ref: '#/components/schemas/UserHomepageStats'
        portfolio:
          type: array
          items:
            $ref: '#/components/schemas/PortfolioItem'
    Notification:
      type: object
      properties:
        id:
          type: integer
          format: int64
        notificationType:
          type: string
        title:
          type: string
        content:
          type: string
        bizType:
          type: string
        bizId:
          type: integer
          format: int64
        readStatus:
          type: string
          enum: [unread, read]
        createdAt:
          type: string
          format: date-time
        readAt:
          type: string
          format: date-time
    ReadNotificationsRequest:
      type: object
      required: [notificationIds]
      properties:
        notificationIds:
          type: array
          items:
            type: integer
            format: int64
...
    PagedTaskSummary:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: '#/components/schemas/TaskSummary'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
...
    PagedNotification:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer